# Conflicts:
#	pom.xml
#	src/main/java/com/ducktem/web/controller/ProductController.java
#	src/main/java/com/ducktem/web/dao/ProductPreviewDao.java
#	src/main/java/com/ducktem/web/service/ProductPreviewService.java
#	src/main/java/com/ducktem/web/service/ProductPreviewServiceImpl.java
#	src/main/resources/Mapper/ProductPreviewDaoMapper.xml
#	src/main/resources/static/js/header.js
#	src/main/resources/templates/detail.html

+++ src/main/java/com/ducktem/web/controller/WishListController.java
@@ -1,17 +1,25 @@
 package com.ducktem.web.controller;
 
+import java.util.ArrayList;
+import java.util.List;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.ducktem.web.entity.Product;
+import com.ducktem.web.entity.ProductPreview;
+import com.ducktem.web.entity.WishList;
 import com.ducktem.web.service.ImgService;
 import com.ducktem.web.service.MemberService;
 import com.ducktem.web.service.ProductService;
@@ -34,23 +42,22 @@ public class WishListController {
     
     @Autowired
     private MemberService memberService;
-//	찜 불러오기
-    @GetMapping("{id}/wish")
-	public String findWishList(HttpSession session,
-								@PathVariable("id") Long productId) {
-		String userId = (String)session.getAttribute("userId");
-		if(userId == null) {
-            System.out.println("로그인 후 사용해주세요.");
-	        return "redirect:/login?returnURL=/product/"+productId+"/wish";
-		}
-		else {		
-			wishListService.findWish(userId, productId);
-	        return "redirect:/";
-		}
+//	나의 찜 목록 불러오기
+    @GetMapping("wish")
+	@ResponseBody
+	public List<WishList> findWishList(HttpSession session) {
+		
+    	String userId = (String)session.getAttribute("userId");
+    	List<WishList> wishList = null;
+		if(userId != null)
+			wishList = wishListService.findWish(userId); 
+		
+		
+		return wishList;
 	}
 	
 //	찜 저장
-    @PostMapping("product/{id}/wish")
+    @PostMapping("{id}/wish")
 	@ResponseBody
 	public String checkWishList(HttpSession session,
 								@PathVariable("id") Long productId) {
@@ -58,10 +65,9 @@ public class WishListController {
         
 		if(userId == null) {
             System.out.println("로그인 후 사용해주세요.");
-            return "reject";
+            return "/login";
 		}
 		else {
-            System.out.println(productId);
 			wishListService.checkWish(userId, productId);
 	        return "ok";
 		}
@@ -70,34 +76,56 @@ public class WishListController {
     
     
 //	찜 취소
-    @PutMapping("{id}/wish")
+    @DeleteMapping("{id}/wish/delete")
 	@ResponseBody
-	public Boolean deleteWishList(HttpSession session,
-								Long productId) {
+	public String deleteWishList(HttpSession session,
+								  @PathVariable("id") Long productId) {
 		String userId = (String)session.getAttribute("userId");
 
 		if(userId == null)
-			return false;
-		else {		
+			return "NoUserId";		
+//			wish 목록에 있는지 확인
+		if(wishListService.confirmWish(userId, productId) == false)	
+			return "NoWishList";
+			
+//			찜 목록 확인	
+		else {			
+			System.out.println(userId);
+			System.out.println(productId);
 	        wishListService.deleteWish(userId, productId);
-	        return true;
+	        return "deleteDone";
 		}
 	}
     
-    
-    
-    
-    
-    
-    
-    
-//	멤버	아이디로 불러오기
-	
-	
-	
+
 	
 //	프로덕트 아이디로 불러오기
-	
-	
+    @GetMapping("{id}/nums")
+	@ResponseBody
+	public int productWishNums(@PathVariable("id") Long productId) {
+    	return wishListService.getWIshNumsbyPId(productId);
+	}
+
+    
+
+//	멤버	아이디로 불러오기 wish 프로덕트 불러오기 
+    @GetMapping("myWishList")
+	@ResponseBody
+	public ArrayList<ProductPreview> myWishList(HttpSession session) {
+    	String userId = (String)session.getAttribute("userId");
+    	List<WishList> wishList = null;    	
+		if(userId != null)
+			wishList = wishListService.findWish(userId); 
+    	return wishListService.getmyWishList(wishList);
+	}
+    
+    
+//	나의 찜 개수 불러오기
+    @GetMapping("myPageWishNum")
+	@ResponseBody
+	public int findMyWishNum(HttpSession session) {
+    	String userId = (String)session.getAttribute("userId");
+		return wishListService.getMyWishNum(userId);
+	}	
 	
 }
